% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogue-utils.R
\name{phi_uniform}
\alias{phi_uniform}
\title{Uniform similarity kernel}
\usage{
phi_uniform(y)
}
\arguments{
\item{y}{numeric vector of length t, the observed time series}
}
\value{
numeric vector of ones with length \code{t - 1}
}
\description{
Returns equal similarity for all previous timepoints.
}
