% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogue-utils.R
\name{return_analogue_preds}
\alias{return_analogue_preds}
\title{Run method of analogues to obtain predictions}
\usage{
return_analogue_preds(y, h, k, method, ...)
}
\arguments{
\item{y}{vector of length t, the observed time series}

\item{h}{integer horizon for which the prediction is desired}

\item{k}{integer number of analogues to use}

\item{method}{which method to use to compute simlarity, one of "uniform",
"seasonal" or "distance"}

\item{...}{other parameters to pass to similarity functions}
}
\value{
named list
}
\description{
Run method of analogues to obtain predictions
}
