% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analogue-utils.R
\name{get_analogues}
\alias{get_analogues}
\title{Find indices of analogues}
\usage{
get_analogues(phi, k = NULL, h)
}
\arguments{
\item{phi}{vector of similarities. Similarities are presumed to be
calculated between observations at timepoint t and i where i is the index
of the vector phi.}

\item{k}{the number of analogues to return. If \code{NULL}, all available
observations are used (useful for marginal/uniform models).}

\item{h}{the horizon to predict for}
}
\value{
a data.frame with three columns.
\itemize{
\item topk_indices: the indices of the original phi vector returned
\item topk_values: the values of the original phi vector
\item topk_weights: the scaled weights (sum to 1) of the phi vector
}
}
\description{
Find indices of analogues
}
