% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analogue_simulation.R
\name{run_analogue_simulation}
\alias{run_analogue_simulation}
\title{Run simulation using method of analogues to determine predictability}
\usage{
run_analogue_simulation(
  data,
  outcome_col,
  h_vals,
  start_idx,
  k_vals_dist,
  k_val_seas
)
}
\arguments{
\item{data}{data.frame with the data. This function assumes that this object
represents a set of time-series observations for a single unit/location. The
data must have:
\itemize{
\item a column named \code{outcome_col} with the outcome in it
\item a column named \code{date} with a YYYY-MM-DD value that can be interpreted as the date of the observation
\item a column named \code{season_week} with the week of the season that the observation was made in
\item a column named \code{season} with a label for the season that the observation was made in
}}

\item{outcome_col}{string with the column name that contains the outcome variable}

\item{h_vals}{vector of integer values for horizons}

\item{start_idx}{integer value for the first row in the dataset to compute
predictions for}

\item{k_vals_dist}{vector of integer values for the number of nearest
neighbours to use in distance-based analogue simulation}

\item{k_val_seas}{integer value for the number of nearest neighbours to use
in seasonal analogue simulation}
}
\value{
named list
}
\description{
Run simulation using method of analogues to determine predictability
}
